misc.c: silence compiler warning
```
C:\projects\libssh2\src\misc.c(83): warning C4018: '<' : signed/unsigned mismatch [C:\projects\libssh2\_builds\src\libssh2.vcxproj]
C:\projects\libssh2\src\misc.c(83): warning C4018: '<' : signed/unsigned mismatch [C:\projects\libssh2\_builds\tests\test_keyboard_interactive_auth_info_request.vcxproj]
```

+++ src/misc.c
@@ -79,7 +79,7 @@ int _libssh2_snprintf(char *cp, size_t cp_max_len, const char *fmt, ...)
     va_start(args, fmt);
     n = vsnprintf(cp, cp_max_len, fmt, args);
     va_end(args);
-    return (n < cp_max_len) ? n : (cp_max_len - 1);
+    return (n < (int)cp_max_len) ? n : (cp_max_len - 1);
 }
 #endif
